fixup! SDL2: preliminary handling SDL_TEXTINPUT
Make SDL2's textinput API more robust. We always send keyboard events if
textinput has not been enabled by the application. Moreover, we only
send key presses as text if the codepoint is valid and if it's not a known
non-printable. Keyboard events are omitted if key presses are sent as
text.

This has been tested with sdl_vnc.

genodelabs/genode-world#368